+++ src/Instrumentum_sanae_doctrinae/web_scraping/monergism/topic/mn_scrap_topic_works.py
@@ -11,13 +11,6 @@ from Instrumentum_sanae_doctrinae.my_tools import general_tools as _my_tools
 from Instrumentum_sanae_doctrinae.web_scraping.monergism.mn_scrap_subtopic_work import MN_ScrapScriptureOrTopicWork, get_subtopics
 
 
-
-class MN_ScrapTopicWork(MN_ScrapScriptureOrTopicWork):
-    def  __init__(self, name, root_folder, url_list, browse_by_type, intermdiate_folders=None):
-        super().__init__(name, root_folder, url_list, browse_by_type, intermdiate_folders)
-        
-
-
 class MN_ScrapTopicWork_All(http_connexion.ParallelHttpConnexionWithLogManagement):
     def __init__(self,root_folder,browse_by_type, overwrite_log=False, update_log=True,intermdiate_folders=None):
 
@@ -84,37 +77,7 @@ class MN_ScrapTopicWork_All(http_connexion.ParallelHttpConnexionWithLogManagemen
         
 
         return input_json_files
-    
-    def prepare_input_data(self,**kwargs):
-        """
-        This method take a json file content and create input data for download 
-        that are put into the dict self.element_dict
-
-        :param file_content: the content of a json file where input data will be taken 
-        :param intermediate_folders: The intermediate folders from the root folder to 
-        the json file 
-        :param file_path: The path of the json file 
-        """
-
-        element = kwargs.get("file_content").get("data")
-        
-        if not element.get("name") in self.element_dict.keys():
-            self.element_dict[element.get("name")]  = []
-        
-        self.element_dict[element.get("name")].append({
-            **{
-                "data":{
-                    "name":element.get("name"),
-                    "pages":element.get("pages")
-                }
-            },
-            **{"download_log":{
-                "input_file_index":self.meta_informations["input_files_information"]\
-                                                        ["input_files"].index(kwargs.get("file_path")),
-                "intermediate_folders":[]} #kwargs.get("intermediate_folders")
-                }
-                })
-        
+            
 
     async def download_element_data(self,element_list):
         """This element take an element ( for example the information of an author or topic) 
@@ -125,7 +88,7 @@ class MN_ScrapTopicWork_All(http_connexion.ParallelHttpConnexionWithLogManagemen
         #print(element.get("data"))
         for element in element_list:
             
-            print(element.get("data").get("name"))
+            #print(element.get("data").get("name"))
     
             ob = MN_ScrapTopicWork(
                 name = element.get("data").get("name"),
