+++ tests/monergism/scripture/test_mn_scrap_scripture_work.py
@@ -5,6 +5,7 @@ import sys
 import logging
 logging.basicConfig(level=logging.DEBUG)
 
+from Instrumentum_sanae_doctrinae.web_scraping.monergism.mn_scrap_work_base import MN_ScrapSpeakerTopicScriptureWork_All
 from Instrumentum_sanae_doctrinae.web_scraping.monergism.scripture import mn_scrap_scripture_works
 
 root_folder = os.path.join(os.getcwd(),'test_folder')
@@ -13,18 +14,20 @@ def test_scrap_all_scripture_work():
     
     browse_by_type = "scripture"
      
-    ob = mn_scrap_scripture_works.MN_ScriptureWork_All(
+    ob = MN_ScrapSpeakerTopicScriptureWork_All(
         root_folder=root_folder,
         browse_by_type=browse_by_type,
-        overwrite_log=True,
+        overwrite_log=False,
     )
+    print(ob.element_dict.keys())
+    
     #print(ob.__dict__)
     
     async def run_ob():
         await ob.print_download_informations(check_from_file=True)
         await ob.download(1)
     
-    asyncio.run(run_ob())
+    #asyncio.run(run_ob())
     
 
 if __name__ == "__main__":
