+++ tests/monergism/topic/test_mn_scrap_topic_work.py
@@ -9,6 +9,7 @@ import unittest
 import logging
 logging.basicConfig(level=logging.DEBUG)
 
+from Instrumentum_sanae_doctrinae.web_scraping.monergism.mn_scrap_work_base import MN_ScrapSpeakerTopicScriptureWork_All
 from Instrumentum_sanae_doctrinae.web_scraping.monergism.topic import mn_scrap_topic_works
 
 
@@ -18,13 +19,13 @@ root_folder = os.path.join(os.getcwd(),'test_folder')
 def test_scrap_all_topic_work():
     browse_by_type = "topic"
      
-    ob = mn_scrap_topic_works.MN_ScrapTopicWork_All(
+    ob = MN_ScrapSpeakerTopicScriptureWork_All(
         root_folder=root_folder,
         browse_by_type=browse_by_type,
-        overwrite_log=True,
+        overwrite_log=False,
     )
-    #print(ob.__dict__)
-    asyncio.run(ob.download(1))
+    print(ob.__dict__)
+    #asyncio.run(ob.download(1))
     
     
 if __name__ == "__main__":
